refactor: 전역 state 제거, main.ts를 조립 전용으로 축소
- Notifier 클래스화 (simple-notify 래핑)
- AppShell / dom/selector.ts로 유틸 이관
- utils/ 폴더 제거, types/state.ts·dom.ts 삭제
