+++ src/main.ts
@@ -1,17 +1,19 @@
 import { SearchForm } from "./search/SearchForm";
-import { getAppElements } from "./utils/AppElementUtil";
-import { notifyEmptyQuery, notifyError } from "./utils/NotifyUtil";
 import { TmdbClient } from "./api/TmdbClient";
 import { MovieListStore } from "./movie-list/MovieListStore";
 import { MovieListView } from "./movie-list/MovieListView";
 import { MovieListController } from "./movie-list/MovieListController";
 import { HeroSection } from "./hero/HeroSection";
-
-const tmdb = new TmdbClient(import.meta.env.VITE_TMDB_API_KEY);
-const movieListStore = new MovieListStore(tmdb);
+import { queryAppShell } from "./dom/AppShell";
+import { Notifier } from "./notify/Notifier";
 
 const main = async () => {
-  const elements = getAppElements();
+  const elements = queryAppShell();
+
+  const notifier = new Notifier();
+
+  const tmdb = new TmdbClient(import.meta.env.VITE_TMDB_API_KEY);
+  const movieListStore = new MovieListStore(tmdb);
 
   const movieListView = new MovieListView({
     listElement: elements.movieList,
@@ -30,7 +32,7 @@ const main = async () => {
     rateValue: elements.heroRateValue,
   });
 
-  const controller = new MovieListController(movieListStore, movieListView, heroSection, { error: notifyError });
+  const controller = new MovieListController(movieListStore, movieListView, heroSection, notifier);
 
   const searchForm = new SearchForm(
     elements.searchForm,
@@ -39,7 +41,7 @@ const main = async () => {
       await controller.search(query);
     },
     () => {
-      notifyEmptyQuery();
+      notifier.warn("검색어를 입력해주세요", "영화 제목을 입력한 뒤 다시 시도해주세요.");
     },
   );
 
@@ -53,5 +55,5 @@ const main = async () => {
 };
 
 window.addEventListener("load", () => {
-  void main().catch((error) => notifyError(error));
+  void main().catch((error) => console.error("[bootstrap failed]", error));
 });
