--- types/dom.ts
@@ -1,17 +1,0 @@
-export interface AppElements {
-  movieList: HTMLUListElement;
-  siteHeader: HTMLElement;
-  searchForm: HTMLFormElement;
-  searchInput: HTMLInputElement;
-  noResult: HTMLDivElement;
-  movieSectionTitle: HTMLHeadingElement;
-
-  heroSection: HTMLElement;
-  heroBackdrop: HTMLDivElement;
-  heroRate: HTMLDivElement;
-  heroRateValue: HTMLSpanElement;
-  heroTitle: HTMLHeadingElement;
-
-  skeletonCard: HTMLUListElement;
-  seeMoreBtn: HTMLButtonElement;
-}
