--- types/state.ts
@@ -1,8 +1,0 @@
-import type { Movie } from "./movie";
-
-export interface State {
-  currentPage: number;
-  totalPage: number;
-  movieList: Movie[];
-  query: string;
-}
